Core/Spells: Make sure that spells with overridden duration but default duration 0 are not flagged immune
Port From (https://github.com/TrinityCore/TrinityCore/commit/ee0c30a968142f940faf64c081babfaae2f4dd18)

+++ Source/Game/Handlers/CalendarHandler.cs
@@ -60,7 +60,7 @@ namespace Game
             {
                 CalendarSendCalendarRaidLockoutInfo lockoutInfo = new();
                 lockoutInfo.MapID = (int)instanceLock.GetMapId();
-                lockoutInfo.DifficultyID = (uint)instanceLock.GetDifficultyId();
+                lockoutInfo.DifficultyID = (short)instanceLock.GetDifficultyId();
                 lockoutInfo.ExpireTime = (int)Math.Max((instanceLock.GetEffectiveExpiryTime() - GameTime.GetSystemTime()).TotalSeconds, 0);
                 lockoutInfo.InstanceID = instanceLock.GetInstanceId();
